Split up HomeScreen into files.
