+++ src/components/HomeScreen/HomeInfo.scss
